Remove unecessary io_mappers
